requirements.txt
grpcio~=1.70.0
protobuf~=5.29.3
google~=3.0.0
aiohttp~=3.11.12